Remove custom forks for e2e tests and refactor helper functions
