+++ plutus-e2e-tests/plutus-e2e-tests.cabal
@@ -69,6 +69,7 @@ common lang
     , mwc-random
     , optparse-applicative
     , prettyprinter
+    , process
     , serialise
     , sqlite-simple
     , stm
@@ -76,6 +77,7 @@ common lang
     , text
     , time
     , transformers-except
+    , unordered-containers
 
 test-suite plutus-e2e-tests-test
   import:         lang
@@ -84,7 +86,12 @@ test-suite plutus-e2e-tests-test
   main-is:        Spec.hs
   hs-source-dirs: test
   other-modules:
-    Helpers
+    CardanoTestnet
+    Helpers.Common
+    Helpers.Query
+    Helpers.Testnet
+    Helpers.Tx
+    Helpers.Utils
     PlutusScripts.AlwaysSucceeds
     PlutusScripts.Helpers
     PlutusScripts.SECP256k1
