+++ plutus-e2e-tests/test/Spec/AlonzoFeatures.hs
@@ -9,13 +9,17 @@
 module Spec.AlonzoFeatures (tests) where
 
 import Cardano.Api qualified as C
+import CardanoTestnet qualified as TN
 import Control.Monad.IO.Class (MonadIO (liftIO))
 import Data.Map qualified as Map
 import Data.Time.Clock.POSIX qualified as Time
 import Hedgehog qualified as H
 import Hedgehog.Extras.Test qualified as HE
-import Helpers (testnetOptionsAlonzo6, testnetOptionsBabbage7, testnetOptionsBabbage8)
-import Helpers qualified as TN
+import Helpers.Query qualified as Q
+import Helpers.Testnet (testnetOptionsAlonzo6, testnetOptionsBabbage7, testnetOptionsBabbage8)
+import Helpers.Testnet qualified as TN
+import Helpers.Tx qualified as Tx
+import Helpers.Utils qualified as U (workspace)
 import Plutus.V1.Ledger.Api qualified as PlutusV1
 import Plutus.V1.Ledger.Interval qualified as PlutusV1
 import Plutus.V1.Ledger.Time qualified as PlutusV1
@@ -25,7 +29,6 @@ import PlutusScripts.V1TxInfo (checkV1TxInfoAssetIdV1, checkV1TxInfoMintWitnessV
 import Test.Base qualified as H
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.Hedgehog (testProperty)
-import Testnet.Plutus qualified as TN
 
 -- | Convert a 'POSIXTime' to the number of milliseconds since the Unix epoch.
 posixToMilliseconds :: Time.POSIXTime -> Integer
@@ -35,13 +38,13 @@ tests :: TestTree
 tests =
   testGroup
     "Alonzo Features"
-    [ testProperty "check each attribute of V1 TxInfo in Alonzo PV6" (checkTxInfoV1 testnetOptionsAlonzo6)
-    , testProperty "check each attribute of V1 TxInfo in Babbage PV7" (checkTxInfoV1 testnetOptionsBabbage7)
-    , testProperty "check each attribute of V1 TxInfo in Babbage PV8" (checkTxInfoV1 testnetOptionsBabbage8)
+    [ testProperty "check each attribute of V1 TxInfo in Alonzo PV6" (checkTxInfoV1Test testnetOptionsAlonzo6)
+    , testProperty "check each attribute of V1 TxInfo in Babbage PV7" (checkTxInfoV1Test testnetOptionsBabbage7)
+    , testProperty "check each attribute of V1 TxInfo in Babbage PV8" (checkTxInfoV1Test testnetOptionsBabbage8)
     ]
 
-checkTxInfoV1 :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
-checkTxInfoV1 networkOptions = H.integration . HE.runFinallies . TN.workspace "." $ \tempAbsPath -> do
+checkTxInfoV1Test :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
+checkTxInfoV1Test networkOptions = H.integration . HE.runFinallies . U.workspace "." $ \tempAbsPath -> do
 
   C.AnyCardanoEra era <- TN.eraFromOptions networkOptions
   preTestnetTime <- liftIO Time.getPOSIXTime
@@ -53,21 +56,21 @@ checkTxInfoV1 networkOptions = H.integration . HE.runFinallies . TN.workspace ".
 
   -- 2: build a transaction
 
-  txIn <- TN.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
-  txInAsTxOut@(C.TxOut _ txInValue _ _) <- TN.getTxOutAtAddress era localNodeConnectInfo w1Address txIn "txInAsTxOut <- TN.getTxOutAtAddress"
+  txIn <- Q.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
+  txInAsTxOut@(C.TxOut _ txInValue _ _) <- Q.getTxOutAtAddress era localNodeConnectInfo w1Address txIn "txInAsTxOut <- TN.getTxOutAtAddress"
 
   let
     tokenValues = C.valueFromList [(checkV1TxInfoAssetIdV1, 1)]
     executionUnits = C.ExecutionUnits {C.executionSteps = 1_000_000_000, C.executionMemory = 10_000_000 }
-    collateral = TN.txInsCollateral era [txIn]
+    collateral = Tx.txInsCollateral era [txIn]
     totalLovelace = C.txOutValueToLovelace txInValue
     fee = 2_000_000 :: C.Lovelace
     amountPaid = 10_000_000
     amountReturned = totalLovelace - amountPaid - fee
     datum = toScriptData (42 ::Integer)
 
-    txOut1 = TN.txOutWithDatumInTx era (C.lovelaceToValue amountPaid <> tokenValues) w1Address datum
-    txOut2 = TN.txOut era (C.lovelaceToValue amountReturned) w1Address
+    txOut1 = Tx.txOutWithDatumInTx era (C.lovelaceToValue amountPaid <> tokenValues) w1Address datum
+    txOut2 = Tx.txOut era (C.lovelaceToValue amountReturned) w1Address
 
     lowerBound = PlutusV1.fromMilliSeconds
       $ PlutusV1.DiffMilliSeconds $ posixToMilliseconds preTestnetTime -- before slot 1
@@ -89,27 +92,26 @@ checkTxInfoV1 networkOptions = H.integration . HE.runFinallies . TN.workspace ".
                expDCert expWdrl expTxInfoValidRange expTxInfoSigs expTxInfoData
     mintWitnesses = Map.fromList [checkV1TxInfoMintWitnessV1 era redeemer executionUnits]
 
-    txBodyContent = (TN.emptyTxBodyContent era pparams)
-      { C.txIns = TN.pubkeyTxIns [txIn]
+    txBodyContent = (Tx.emptyTxBodyContent era pparams)
+      { C.txIns = Tx.pubkeyTxIns [txIn]
       , C.txInsCollateral = collateral
-      , C.txMintValue = TN.txMintValue era tokenValues mintWitnesses
+      , C.txMintValue = Tx.txMintValue era tokenValues mintWitnesses
       , C.txOuts = [txOut1, txOut2]
-      , C.txFee = TN.txFee era fee
-      , C.txValidityRange = TN.txValidityRange era 1 2700 -- ~9min range (200ms slots). Babbage era onwards cannot have upper slot beyond epoch boundary (10_000 slot epoch).
-      , C.txExtraKeyWits = TN.txExtraKeyWits era [w1VKey]
+      , C.txFee = Tx.txFee era fee
+      , C.txValidityRange = Tx.txValidityRange era 1 2700 -- ~9min range (200ms slots). Babbage era onwards cannot have upper slot beyond epoch boundary (10_000 slot epoch).
+      , C.txExtraKeyWits = Tx.txExtraKeyWits era [w1VKey]
       }
-  txbody <- TN.buildRawTx era txBodyContent
-  kw <- TN.signTx era txbody w1SKey
+  txbody <- Tx.buildRawTx era txBodyContent
+  kw <- Tx.signTx era txbody w1SKey
   let signedTx = C.makeSignedTransaction [kw] txbody
 
-  TN.submitTx era localNodeConnectInfo signedTx
+  Tx.submitTx era localNodeConnectInfo signedTx
 
-  let expectedTxIn = TN.txIn (TN.txId signedTx) 0
-  resultTxOut <- TN.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "resultTxOut <- TN.getTxOutAtAddress "
-  txOutHasTokenValue <- TN.txOutHasValue resultTxOut tokenValues
+  let expectedTxIn = Tx.txIn (Tx.txId signedTx) 0
+  resultTxOut <- Q.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "resultTxOut <- TN.getTxOutAtAddress "
+  txOutHasTokenValue <- Q.txOutHasValue resultTxOut tokenValues
   H.assert txOutHasTokenValue
   H.success
 
 -- TODO: datumHashSpendTest
 -- TODO: mintTest
--- TODO: can't use V2 script test
