+++ plutus-e2e-tests/test/Spec/BabbageFeatures.hs
@@ -16,31 +16,35 @@ import Hedgehog.Extras.Test qualified as HE
 import Test.Base qualified as H
 import Test.Tasty.Hedgehog (testProperty)
 
-import Helpers (testnetOptionsBabbage7, testnetOptionsBabbage8)
-import Helpers qualified as TN
+import CardanoTestnet qualified as TN
+import Helpers.Common (makeAddress)
+import Helpers.Query qualified as Q
+import Helpers.Testnet (testnetOptionsBabbage7, testnetOptionsBabbage8)
+import Helpers.Testnet qualified as TN
+import Helpers.Tx qualified as Tx
+import Helpers.Utils qualified as U (workspace)
 import PlutusScripts.AlwaysSucceeds qualified as PS
 import PlutusScripts.Helpers qualified as PS
-import Testnet.Plutus qualified as TN
 
 tests :: TestTree
 tests = testGroup "Babbage Features"
   [ testGroup "reference script"
-    [ testProperty "mint a token with a reference script in Babbage PV7" (referenceScriptMint testnetOptionsBabbage7)
-    , testProperty "mint a token with a reference script in Babbage PV8" (referenceScriptMint testnetOptionsBabbage8)
-    --, testProperty "mint a token with a reference script in Babbage PV8" (referenceScriptMint localNodeOptionsPreview) -- uncomment to use local node on preview testnet
+    [ testProperty "mint a token with a reference script in Babbage PV7" (referenceScriptMintTest testnetOptionsBabbage7)
+    , testProperty "mint a token with a reference script in Babbage PV8" (referenceScriptMintTest testnetOptionsBabbage8)
+    --, testProperty "mint a token with a reference script in Babbage PV8" (referenceScriptMintTest localNodeOptionsPreview) -- uncomment to use local node on preview testnet
 
-    , testProperty "spend locked funds with a reference script using inline datum in Babbage PV7" (referenceScriptInlineDatumSpend testnetOptionsBabbage7)
-    , testProperty "spend locked funds with a reference script using inline datum in Babbage PV8" (referenceScriptInlineDatumSpend testnetOptionsBabbage8)
-    --, testProperty "spend locked funds with a reference script using inline datum in Babbage PV8" (referenceScriptInlineDatumSpend localNodeOptionsPreview) -- uncomment to use local node on preview testnet
+    , testProperty "spend locked funds with a reference script using inline datum in Babbage PV7" (referenceScriptInlineDatumSpendTest testnetOptionsBabbage7)
+    , testProperty "spend locked funds with a reference script using inline datum in Babbage PV8" (referenceScriptInlineDatumSpendTest testnetOptionsBabbage8)
+    --, testProperty "spend locked funds with a reference script using inline datum in Babbage PV8" (referenceScriptInlineDatumSpendTest localNodeOptionsPreview) -- uncomment to use local node on preview testnet
 
-    , testProperty "spend locked funds with a reference script providing datum in txbody in Babbage PV7" (referenceScriptDatumHashSpend testnetOptionsBabbage7)
-    , testProperty "spend locked funds with a reference script providing datum in txbody in Babbage PV8" (referenceScriptDatumHashSpend testnetOptionsBabbage8)
-    --, testProperty "spend locked funds with a reference script providing datum in txbody in Babbage PV8" (referenceScriptDatumHashSpend localNodeOptionsPreview) -- uncomment to use local node on preview testnet
+    , testProperty "spend locked funds with a reference script providing datum in txbody in Babbage PV7" (referenceScriptDatumHashSpendTest testnetOptionsBabbage7)
+    , testProperty "spend locked funds with a reference script providing datum in txbody in Babbage PV8" (referenceScriptDatumHashSpendTest testnetOptionsBabbage8)
+    --, testProperty "spend locked funds with a reference script providing datum in txbody in Babbage PV8" (referenceScriptDatumHashSpendTest localNodeOptionsPreview) -- uncomment to use local node on preview testnet
     ]
   ]
 
-referenceScriptMint :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
-referenceScriptMint networkOptions = H.integration . HE.runFinallies . TN.workspace "." $ \tempAbsPath -> do
+referenceScriptMintTest :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
+referenceScriptMintTest networkOptions = H.integration . HE.runFinallies . U.workspace "." $ \tempAbsPath -> do
 
   C.AnyCardanoEra era <- TN.eraFromOptions networkOptions
 
@@ -50,52 +54,52 @@ referenceScriptMint networkOptions = H.integration . HE.runFinallies . TN.worksp
 
   -- 2: build a transaction to hold reference script
 
-  txIn <- TN.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
+  txIn <- Q.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
 
   let
-    refScriptTxOut = TN.txOutWithRefScript era (C.lovelaceToValue 20_000_000) w1Address
+    refScriptTxOut = Tx.txOutWithRefScript era (C.lovelaceToValue 20_000_000) w1Address
       (PS.unPlutusScriptV2 PS.alwaysSucceedPolicyScriptV2)
-    otherTxOut = TN.txOut era (C.lovelaceToValue 5_000_000) w1Address
+    otherTxOut = Tx.txOut era (C.lovelaceToValue 5_000_000) w1Address
 
-    txBodyContent = (TN.emptyTxBodyContent era pparams)
-      { C.txIns = TN.pubkeyTxIns [txIn]
+    txBodyContent = (Tx.emptyTxBodyContent era pparams)
+      { C.txIns = Tx.pubkeyTxIns [txIn]
       , C.txOuts = [refScriptTxOut, otherTxOut]
       }
 
-  signedTx <- TN.buildTx era txBodyContent w1Address w1SKey networkId
-  TN.submitTx era localNodeConnectInfo signedTx
-  let refScriptTxIn = TN.txIn (TN.txId signedTx) 0
-      otherTxIn   = TN.txIn (TN.txId signedTx) 1
-  TN.waitForTxInAtAddress era localNodeConnectInfo w1Address refScriptTxIn "TN.waitForTxInAtAddress"
+  signedTx <- Tx.buildTx era txBodyContent w1Address w1SKey networkId
+  Tx.submitTx era localNodeConnectInfo signedTx
+  let refScriptTxIn = Tx.txIn (Tx.txId signedTx) 0
+      otherTxIn   = Tx.txIn (Tx.txId signedTx) 1
+  Q.waitForTxInAtAddress era localNodeConnectInfo w1Address refScriptTxIn "Tx.waitForTxInAtAddress"
 
   -- 3: build a transaction to mint token using reference script
 
   let
     tokenValues = C.valueFromList [(PS.alwaysSucceedAssetIdV2, 6)]
     mintWitnesses = Map.fromList [PS.alwaysSucceedMintWitnessV2 era (Just refScriptTxIn)]
-    collateral = TN.txInsCollateral era [otherTxIn]
-    txOut = TN.txOut era (C.lovelaceToValue 3_000_000 <> tokenValues) w1Address
+    collateral = Tx.txInsCollateral era [otherTxIn]
+    txOut = Tx.txOut era (C.lovelaceToValue 3_000_000 <> tokenValues) w1Address
 
-    txBodyContent2 = (TN.emptyTxBodyContent era pparams)
-      { C.txIns = TN.pubkeyTxIns [otherTxIn]
+    txBodyContent2 = (Tx.emptyTxBodyContent era pparams)
+      { C.txIns = Tx.pubkeyTxIns [otherTxIn]
       , C.txInsCollateral = collateral
-      , C.txInsReference = TN.txInsReference era [refScriptTxIn]
-      , C.txMintValue = TN.txMintValue era tokenValues mintWitnesses
+      , C.txInsReference = Tx.txInsReference era [refScriptTxIn]
+      , C.txMintValue = Tx.txMintValue era tokenValues mintWitnesses
       , C.txOuts = [txOut]
       }
 
-  signedTx2 <- TN.buildTx era txBodyContent2 w1Address w1SKey networkId
-  TN.submitTx era localNodeConnectInfo signedTx2
-  let expectedTxIn = TN.txIn (TN.txId signedTx2) 0
+  signedTx2 <- Tx.buildTx era txBodyContent2 w1Address w1SKey networkId
+  Tx.submitTx era localNodeConnectInfo signedTx2
+  let expectedTxIn = Tx.txIn (Tx.txId signedTx2) 0
   -- Query for txo and assert it contains newly minted token
-  resultTxOut <- TN.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "TN.getTxOutAtAddress"
-  txOutHasTokenValue <- TN.txOutHasValue resultTxOut tokenValues
+  resultTxOut <- Q.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "Tx.getTxOutAtAddress"
+  txOutHasTokenValue <- Q.txOutHasValue resultTxOut tokenValues
   H.assert txOutHasTokenValue
   H.success
 
 
-referenceScriptInlineDatumSpend :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
-referenceScriptInlineDatumSpend networkOptions = H.integration . HE.runFinallies . TN.workspace "." $ \tempAbsPath -> do
+referenceScriptInlineDatumSpendTest :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
+referenceScriptInlineDatumSpendTest networkOptions = H.integration . HE.runFinallies . U.workspace "." $ \tempAbsPath -> do
 
   C.AnyCardanoEra era <- TN.eraFromOptions networkOptions
 
@@ -105,54 +109,54 @@ referenceScriptInlineDatumSpend networkOptions = H.integration . HE.runFinallies
 
   -- 2: build a transaction to hold reference script
 
-  txIn <- TN.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
+  txIn <- Q.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
 
   let
-    refTxOut      = TN.txOutWithRefScript era (C.lovelaceToValue 20_000_000) w1Address
+    refTxOut      = Tx.txOutWithRefScript era (C.lovelaceToValue 20_000_000) w1Address
                      (PS.unPlutusScriptV2 PS.alwaysSucceedSpendScriptV2)
-    otherTxOut    = TN.txOut era (C.lovelaceToValue 5_000_000) w1Address
-    scriptAddress = TN.makeAddress (Right PS.alwaysSucceedSpendScriptHashV2) networkId
-    scriptTxOut   = TN.txOutWithInlineDatum era (C.lovelaceToValue 10_000_000) scriptAddress (PS.toScriptData ())
+    otherTxOut    = Tx.txOut era (C.lovelaceToValue 5_000_000) w1Address
+    scriptAddress = makeAddress (Right PS.alwaysSucceedSpendScriptHashV2) networkId
+    scriptTxOut   = Tx.txOutWithInlineDatum era (C.lovelaceToValue 10_000_000) scriptAddress (PS.toScriptData ())
 
-    txBodyContent = (TN.emptyTxBodyContent era pparams)
-      { C.txIns = TN.pubkeyTxIns [txIn]
+    txBodyContent = (Tx.emptyTxBodyContent era pparams)
+      { C.txIns = Tx.pubkeyTxIns [txIn]
       , C.txOuts = [refTxOut, otherTxOut, scriptTxOut]
       }
 
-  signedTx <- TN.buildTx era txBodyContent w1Address w1SKey networkId
-  TN.submitTx era localNodeConnectInfo signedTx
-  let refScriptTxIn = TN.txIn (TN.txId signedTx) 0
-      otherTxIn     = TN.txIn (TN.txId signedTx) 1
-      txInAtScript  = TN.txIn (TN.txId signedTx) 2
-  TN.waitForTxInAtAddress era localNodeConnectInfo w1Address refScriptTxIn "TN.waitForTxInAtAddress"
+  signedTx <- Tx.buildTx era txBodyContent w1Address w1SKey networkId
+  Tx.submitTx era localNodeConnectInfo signedTx
+  let refScriptTxIn = Tx.txIn (Tx.txId signedTx) 0
+      otherTxIn     = Tx.txIn (Tx.txId signedTx) 1
+      txInAtScript  = Tx.txIn (Tx.txId signedTx) 2
+  Q.waitForTxInAtAddress era localNodeConnectInfo w1Address refScriptTxIn "Tx.waitForTxInAtAddress"
 
   -- 3: build a transaction to mint token using reference script
 
   let
-    scriptTxIn = TN.txInWitness txInAtScript (PS.alwaysSucceedSpendWitnessV2 era (Just refScriptTxIn) Nothing)
-    collateral = TN.txInsCollateral era [otherTxIn]
+    scriptTxIn = Tx.txInWitness txInAtScript (PS.alwaysSucceedSpendWitnessV2 era (Just refScriptTxIn) Nothing)
+    collateral = Tx.txInsCollateral era [otherTxIn]
     adaValue = C.lovelaceToValue 4_200_000
-    txOut = TN.txOut era adaValue w1Address
+    txOut = Tx.txOut era adaValue w1Address
 
-    txBodyContent2 = (TN.emptyTxBodyContent era pparams)
+    txBodyContent2 = (Tx.emptyTxBodyContent era pparams)
       { C.txIns = [scriptTxIn]
-      , C.txInsReference = TN.txInsReference era [refScriptTxIn]
+      , C.txInsReference = Tx.txInsReference era [refScriptTxIn]
       , C.txInsCollateral = collateral
       , C.txOuts = [txOut]
       }
 
-  signedTx2 <- TN.buildTx era txBodyContent2 w1Address w1SKey networkId
-  TN.submitTx era localNodeConnectInfo signedTx2
-  let expectedTxIn = TN.txIn (TN.txId signedTx2) 0
+  signedTx2 <- Tx.buildTx era txBodyContent2 w1Address w1SKey networkId
+  Tx.submitTx era localNodeConnectInfo signedTx2
+  let expectedTxIn = Tx.txIn (Tx.txId signedTx2) 0
   -- Query for txo and assert it contains newly minted token
-  resultTxOut <- TN.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "TN.getTxOutAtAddress"
-  txOutHasAdaValue <- TN.txOutHasValue resultTxOut adaValue
+  resultTxOut <- Q.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "Tx.getTxOutAtAddress"
+  txOutHasAdaValue <- Q.txOutHasValue resultTxOut adaValue
   H.assert txOutHasAdaValue
   H.success
 
 
-referenceScriptDatumHashSpend :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
-referenceScriptDatumHashSpend networkOptions = H.integration . HE.runFinallies . TN.workspace "." $ \tempAbsPath -> do
+referenceScriptDatumHashSpendTest :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
+referenceScriptDatumHashSpendTest networkOptions = H.integration . HE.runFinallies . U.workspace "." $ \tempAbsPath -> do
 
   C.AnyCardanoEra era <- TN.eraFromOptions networkOptions
 
@@ -162,49 +166,49 @@ referenceScriptDatumHashSpend networkOptions = H.integration . HE.runFinallies .
 
   -- 2: build a transaction to hold reference script
 
-  txIn <- TN.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
+  txIn <- Q.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
 
   let
-    refTxOut      = TN.txOutWithRefScript era (C.lovelaceToValue 20_000_000) w1Address
+    refTxOut      = Tx.txOutWithRefScript era (C.lovelaceToValue 20_000_000) w1Address
                      (PS.unPlutusScriptV2 PS.alwaysSucceedSpendScriptV2)
-    otherTxOut    = TN.txOut era (C.lovelaceToValue 5_000_000) w1Address
-    scriptAddress = TN.makeAddress (Right PS.alwaysSucceedSpendScriptHashV2) networkId
+    otherTxOut    = Tx.txOut era (C.lovelaceToValue 5_000_000) w1Address
+    scriptAddress = makeAddress (Right PS.alwaysSucceedSpendScriptHashV2) networkId
     datum         = PS.toScriptData ()
-    scriptTxOut   = TN.txOutWithDatumHash era (C.lovelaceToValue 10_000_000) scriptAddress datum
+    scriptTxOut   = Tx.txOutWithDatumHash era (C.lovelaceToValue 10_000_000) scriptAddress datum
 
-    txBodyContent = (TN.emptyTxBodyContent era pparams)
-      { C.txIns = TN.pubkeyTxIns [txIn]
+    txBodyContent = (Tx.emptyTxBodyContent era pparams)
+      { C.txIns = Tx.pubkeyTxIns [txIn]
       , C.txOuts = [refTxOut, otherTxOut, scriptTxOut]
       }
 
-  signedTx <- TN.buildTx era txBodyContent w1Address w1SKey networkId
-  TN.submitTx era localNodeConnectInfo signedTx
-  let refScriptTxIn = TN.txIn (TN.txId signedTx) 0
-      otherTxIn     = TN.txIn (TN.txId signedTx) 1
-      txInAtScript  = TN.txIn (TN.txId signedTx) 2
-  TN.waitForTxInAtAddress era localNodeConnectInfo w1Address refScriptTxIn "TN.waitForTxInAtAddress"
+  signedTx <- Tx.buildTx era txBodyContent w1Address w1SKey networkId
+  Tx.submitTx era localNodeConnectInfo signedTx
+  let refScriptTxIn = Tx.txIn (Tx.txId signedTx) 0
+      otherTxIn     = Tx.txIn (Tx.txId signedTx) 1
+      txInAtScript  = Tx.txIn (Tx.txId signedTx) 2
+  Q.waitForTxInAtAddress era localNodeConnectInfo w1Address refScriptTxIn "Tx.waitForTxInAtAddress"
 
   -- 3: build a transaction to mint token using reference script
 
   let
-    scriptTxIn = TN.txInWitness txInAtScript $ PS.alwaysSucceedSpendWitnessV2 era (Just refScriptTxIn) (Just datum)
-    collateral = TN.txInsCollateral era [otherTxIn]
+    scriptTxIn = Tx.txInWitness txInAtScript $ PS.alwaysSucceedSpendWitnessV2 era (Just refScriptTxIn) (Just datum)
+    collateral = Tx.txInsCollateral era [otherTxIn]
     adaValue = C.lovelaceToValue 4_200_000
-    txOut = TN.txOut era adaValue w1Address
+    txOut = Tx.txOut era adaValue w1Address
 
-    txBodyContent2 = (TN.emptyTxBodyContent era pparams)
+    txBodyContent2 = (Tx.emptyTxBodyContent era pparams)
       { C.txIns = [scriptTxIn]
-      , C.txInsReference = TN.txInsReference era [refScriptTxIn]
+      , C.txInsReference = Tx.txInsReference era [refScriptTxIn]
       , C.txInsCollateral = collateral
       , C.txOuts = [txOut]
       }
 
-  signedTx2 <- TN.buildTx era txBodyContent2 w1Address w1SKey networkId
-  TN.submitTx era localNodeConnectInfo signedTx2
-  let expectedTxIn = TN.txIn (TN.txId signedTx2) 0
+  signedTx2 <- Tx.buildTx era txBodyContent2 w1Address w1SKey networkId
+  Tx.submitTx era localNodeConnectInfo signedTx2
+  let expectedTxIn = Tx.txIn (Tx.txId signedTx2) 0
   -- Query for txo and assert it contains newly minted token
-  resultTxOut <- TN.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "TN.getTxOutAtAddress"
-  txOutHasAdaValue <- TN.txOutHasValue resultTxOut adaValue
+  resultTxOut <- Q.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "Tx.getTxOutAtAddress"
+  txOutHasAdaValue <- Q.txOutHasValue resultTxOut adaValue
   H.assert txOutHasAdaValue
   H.success
