+++ plutus-e2e-tests/test/Spec/Builtins/SECP256k1.hs
@@ -18,17 +18,20 @@ import Hedgehog.Extras.Test qualified as HE
 import Test.Base qualified as H
 import Test.Tasty.Hedgehog (testProperty)
 
-import Helpers (testnetOptionsAlonzo6, testnetOptionsBabbage7, testnetOptionsBabbage8)
-import Helpers qualified as TN
+import CardanoTestnet qualified as TN
+import Helpers.Query qualified as Q
+import Helpers.Testnet (testnetOptionsAlonzo6, testnetOptionsBabbage7, testnetOptionsBabbage8)
+import Helpers.Testnet qualified as TN
+import Helpers.Tx qualified as Tx
+import Helpers.Utils qualified as U
 import PlutusScripts.SECP256k1 qualified as PS
-import Testnet.Plutus qualified as TN
 
 tests :: TestTree
 tests = testGroup "SECP256k1"
-  [ testProperty "unable to use SECP256k1 builtins in Alonzo PV6" (verifySchnorrAndEcdsa testnetOptionsAlonzo6)
-  , testProperty "unable to use SECP256k1 builtins in Babbage PV7" (verifySchnorrAndEcdsa testnetOptionsBabbage7)
-  , testProperty "can use SECP256k1 builtins in Babbage PV8" (verifySchnorrAndEcdsa testnetOptionsBabbage8)
-  --, testProperty "can use SECP256k1 builtins in Babbage PV8 (on preview testnet)" (verifySchnorrAndEcdsa localNodeOptionsPreview) -- uncomment to use local node on preview testnet
+  [ testProperty "unable to use SECP256k1 builtins in Alonzo PV6" (verifySchnorrAndEcdsaTest testnetOptionsAlonzo6)
+  , testProperty "unable to use SECP256k1 builtins in Babbage PV7" (verifySchnorrAndEcdsaTest testnetOptionsBabbage7)
+  , testProperty "can use SECP256k1 builtins in Babbage PV8" (verifySchnorrAndEcdsaTest testnetOptionsBabbage8)
+  --, testProperty "can use SECP256k1 builtins in Babbage PV8 (on preview testnet)" (verifySchnorrAndEcdsaTest localNodeOptionsPreview) -- uncomment to use local node on preview testnet
   ]
 
 {- | Test that builtins: verifySchnorrSecp256k1Signature and verifyEcdsaSecp256k1Signature can only
@@ -40,8 +43,8 @@ tests = testGroup "SECP256k1"
     - if pv8+ then query the ledger to see if mint was successful otherwise expect
         "forbidden builtin" error when building tx
 -}
-verifySchnorrAndEcdsa :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
-verifySchnorrAndEcdsa networkOptions = H.integration . HE.runFinallies . TN.workspace "." $ \tempAbsPath -> do
+verifySchnorrAndEcdsaTest :: Either TN.LocalNodeOptions TN.TestnetOptions -> H.Property
+verifySchnorrAndEcdsaTest networkOptions = H.integration . HE.runFinallies . U.workspace "." $ \tempAbsPath -> do
 
   pv <- TN.pvFromOptions networkOptions
   C.AnyCardanoEra era <- TN.eraFromOptions networkOptions
@@ -52,7 +55,7 @@ verifySchnorrAndEcdsa networkOptions = H.integration . HE.runFinallies . TN.work
 
 -- 2: build a transaction
 
-  txIn <- TN.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
+  txIn <- Q.adaOnlyTxInAtAddress era localNodeConnectInfo w1Address
 
   let
     (verifySchnorrAssetId, verifyEcdsaAssetId, verifySchnorrMintWitness, verifyEcdsaMintWitness) =
@@ -69,34 +72,34 @@ verifySchnorrAndEcdsa networkOptions = H.integration . HE.runFinallies . TN.work
             PS.verifyEcdsaMintWitnessV2 era )
 
     tokenValues = C.valueFromList [(verifySchnorrAssetId, 4), (verifyEcdsaAssetId, 2)]
-    txOut = TN.txOut era (C.lovelaceToValue 3_000_000 <> tokenValues) w1Address
+    txOut = Tx.txOut era (C.lovelaceToValue 3_000_000 <> tokenValues) w1Address
     mintWitnesses = Map.fromList [verifySchnorrMintWitness, verifyEcdsaMintWitness]
-    collateral = TN.txInsCollateral era [txIn]
-    txBodyContent = (TN.emptyTxBodyContent era pparams)
-      { C.txIns = TN.pubkeyTxIns [txIn]
+    collateral = Tx.txInsCollateral era [txIn]
+    txBodyContent = (Tx.emptyTxBodyContent era pparams)
+      { C.txIns = Tx.pubkeyTxIns [txIn]
       , C.txInsCollateral = collateral
-      , C.txMintValue = TN.txMintValue era tokenValues mintWitnesses
+      , C.txMintValue = Tx.txMintValue era tokenValues mintWitnesses
       , C.txOuts = [txOut]
       }
 
   case pv < 8 of
     True -> do
       -- Assert that "forbidden" error occurs when attempting to use either SECP256k1 builtin
-      eitherTx <- TN.buildTx' era txBodyContent w1Address w1SKey networkId
-      H.assert $ TN.isTxBodyScriptExecutionError
+      eitherTx <- Tx.buildTx' era txBodyContent w1Address w1SKey networkId
+      H.assert $ Tx.isTxBodyScriptExecutionError
         "Forbidden builtin function: (builtin verifySchnorrSecp256k1Signature)" eitherTx
-      H.assert $ TN.isTxBodyScriptExecutionError
+      H.assert $ Tx.isTxBodyScriptExecutionError
         "Forbidden builtin function: (builtin verifyEcdsaSecp256k1Signature)" eitherTx
       H.success
 
     False -> do
       -- Build and submit transaction
-      signedTx <- TN.buildTx era txBodyContent w1Address w1SKey networkId
-      TN.submitTx era localNodeConnectInfo signedTx
-      let expectedTxIn = TN.txIn (TN.txId signedTx) 0
+      signedTx <- Tx.buildTx era txBodyContent w1Address w1SKey networkId
+      Tx.submitTx era localNodeConnectInfo signedTx
+      let expectedTxIn = Tx.txIn (Tx.txId signedTx) 0
 
-      -- Query for txo and assert it contains newly minting tokens to prove successfuluse of SECP256k1 builtins
-      resultTxOut <- TN.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "TN.getTxOutAtAddress"
-      txOutHasTokenValue <- TN.txOutHasValue resultTxOut tokenValues
+      -- Query for txo and assert it contains newly minting tokens to prove successful use of SECP256k1 builtins
+      resultTxOut <- Q.getTxOutAtAddress era localNodeConnectInfo w1Address expectedTxIn "TN.getTxOutAtAddress"
+      txOutHasTokenValue <- Q.txOutHasValue resultTxOut tokenValues
       H.assert txOutHasTokenValue
       H.success
